# test: Fix race condition in LogQL test